requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
faiss-cpu>=1.12.0
sentence-transformers==5.1.0
openai>=1.16.0
httpx>=0.27.0
sqlalchemy==1.4.52
alembic==1.12.1
pydantic>=2.5.0
pydantic-settings==2.10.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2